Update Build_and_Deploy.md
This is how I was able to build. 

When you change the directory, there is a line in docker script `ADD . /src` It just copies the build-scripts when it needs the source code.
https://github.com/monero-project/monero/issues/3039

So you call docker from root of source then you can directly add it. 

Also the build is in /src/

see the last lines 
```

ADD . /src
RUN cd /src \
    && CMAKE_INCLUDE_PATH="${PREFIX}/include" \
       CMAKE_LIBRARY_PATH="${PREFIX}/lib" \
       ANDROID_STANDALONE_TOOLCHAIN_PATH=${TOOLCHAIN_DIR} \
       USE_SINGLE_BUILDDIR=1 \
       PATH=${HOST_PATH} make release-static-android-armv8 -j${NPROC}
```